Consolidate all navigation buttons into navigation.cs and remove view specific scripts. Change selected season button highlighting style to show brigher when selected.
