--- Assets/Scripts/customer_view_button.cs
@@ -1,12 +1,0 @@
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class customer_view_button : MonoBehaviour
-{
-    
-    public void customer_view()
-    {
-        SceneManager.LoadScene("customer_view");
-    }
-
-}
