--- Assets/Scripts/customer_view_button.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 797b6d55171a20f439ee9f695a36af73
