--- Assets/Scripts/kitchen_button.cs
@@ -1,14 +1,0 @@
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class kitchen_button : MonoBehaviour
-{
-   
-    public void kitchen_view()
-    {
-        SceneManager.LoadScene("kitchen_view");
-    }
-
-
-
-}
