--- Assets/Scripts/kitchen_button.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: dc2e6f8ff86befc46979a68ff3f02dbb
