--- Assets/Scripts/quit_button.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 702e457d6c2422e46a45a9e2151e4f9a
