+++ Assets/kitchen_view.unity
@@ -196,7 +196,6 @@ GameObject:
   - component: {fileID: 147707078}
   - component: {fileID: 147707077}
   - component: {fileID: 147707076}
-  - component: {fileID: 147707075}
   - component: {fileID: 147707080}
   m_Layer: 5
   m_Name: Canvas
@@ -205,18 +204,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &147707075
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 147707074}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 797b6d55171a20f439ee9f695a36af73, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &147707076
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -310,7 +297,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 147707074}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 702e457d6c2422e46a45a9e2151e4f9a, type: 3}
+  m_Script: {fileID: 11500000, guid: 8b1764021081448cfaefd5a1b8251321, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &746335000
@@ -396,7 +383,7 @@ MonoBehaviour:
     m_PersistentCalls:
       m_Calls:
       - m_Target: {fileID: 147707080}
-        m_TargetAssemblyTypeName: quit_button, Assembly-CSharp
+        m_TargetAssemblyTypeName: Navigation, Assembly-CSharp
         m_MethodName: quit_game
         m_Mode: 1
         m_Arguments:
@@ -664,8 +651,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 147707075}
-        m_TargetAssemblyTypeName: customer_view_button, Assembly-CSharp
+      - m_Target: {fileID: 147707080}
+        m_TargetAssemblyTypeName: Navigation, Assembly-CSharp
         m_MethodName: customer_view
         m_Mode: 1
         m_Arguments:
